Updates to allow proxying via the mmhmm service API
In order to use the Deepgram REST API from the browser, we need to proxy
requests through our backend - their backend doesn't support CORS.

When we proxy requests, we want to include our own authentication token
in the Authorization header so that only authenticated mmhmm can access
the proxy endpoint. The Deepgram SDK was prohibiting this; they required
the use of the literal string "proxy" in the Authorization header when
proxying.

When we proxy requests, we want to prepend a path segment so that we can
host this under api.mmhmm.app, instead of having to use a separate
hostname for the proxy. The Deepgram SDK was prohibiting this by
removing the trailing slash from our proxy URL.