fix: support bearer token authorization
The standard format for bearer token authorization is:
Authorization: Bearer {token}

However, Deepgram replaces the word "Bearer" with "Token":
Authorization: Token {token}

When proxying REST requests through our backend, we need to use OUR
authorization mechanism, which requires the standard "Bearer". With this
change we look for the "Bearer " prefix in the API key and, if found,
use it as-is. Otherwise we maintain the existing Deepgram behavior.